Add weighted formula to kingdom generator.
Add simple form to the kingdom generator pop-up that will generate a
string like ragingduckd's bash script in the "Formula for modified card
frequencies" thread at f.ds (linked below).

In the current implementation the user is locked into ragingduckd's
definitions of "more" and "less", but we could change the UI to allow
the user to add any number of fields, setting each of those fields to
any frequency relative to the "normal" frequency that applies to the
rest of the cards.

As the feature grows, we might want to add checks for cards being named
in multiple lists, or too any cards included in an "include" field, and
keep an eye on performance with the Jison-generated code in
kingdom_generator.js; if we allow user-entered relative frequencies,
they will need to be converted to integers and could potentially end up
being fairly large, and I have no idea if that would impact
performance. Probably I am just being paranoid :-)

http://forum.dominionstrategy.com/index.php?topic=11749.0